refactor(templates): break TemplateManagerScreen into sub-components
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
