+++ src/components/templates/templateManagerShared.ts
@@ -1,0 +1,18 @@
+import { NoteTemplateIcon } from '../../services/TemplateService';
+
+export const ICON_OPTIONS: NoteTemplateIcon[] = [
+  'document-outline',
+  'people-outline',
+  'bulb-outline',
+  'code-slash-outline',
+  'book-outline',
+  'bug-outline',
+  'checkmark-done-outline',
+  'calendar-outline',
+  'pencil-outline',
+  'flask-outline',
+  'rocket-outline',
+  'flag-outline',
+];
+
+export const DEFAULT_ICON: NoteTemplateIcon = 'document-outline';
